Connect MySQL and Change id's type to Integer

+++ src/main/java/us/repository/MeetingRepository.java
@@ -3,6 +3,6 @@ package us.repository;
 import org.springframework.data.repository.CrudRepository;
 import us.model.Meeting;
 
-public interface MeetingRepository extends CrudRepository<Meeting, Long>{
+public interface MeetingRepository extends CrudRepository<Meeting, Integer>{
 
 }
