+++ src/main/java/us/repository/ParticipantRepository.java
@@ -3,5 +3,5 @@ package us.repository;
 import org.springframework.data.repository.CrudRepository;
 import us.model.Participant;
 
-public interface ParticipantRepository extends CrudRepository<Participant, Long> {
+public interface ParticipantRepository extends CrudRepository<Participant, Integer> {
 }
